requirements.txt
pyopengl
glfw
numpy